support switch physics backend
