+++ cocos/physics/cocos/instantiate.ts
@@ -35,7 +35,7 @@ import { BuiltinBoxShape } from './shapes/builtin-box-shape';
 import { BuiltinSphereShape } from './shapes/builtin-sphere-shape';
 import { BuiltinCapsuleShape } from './shapes/builtin-capsule-shape';
 
-selector.select('builtin', {
+selector.register('builtin', {
     RigidBody: BuiltinRigidBody,
     BoxShape: BuiltinBoxShape,
     SphereShape: BuiltinSphereShape,
