+++ cocos/physics/framework/components/colliders/collider.ts
@@ -41,7 +41,7 @@ import { Component, error, Node } from '../../../../core';
 import { IBaseShape } from '../../../spec/i-physics-shape';
 import { AABB, Sphere } from '../../../../core/geometry';
 import { EColliderType, EAxisDirection } from '../../physics-enum';
-import { createShape } from '../../instance';
+import { createShape } from '../../physics-selector';
 
 /**
  * @en
