+++ cocos/physics/framework/components/constraints/constraint.ts
@@ -35,7 +35,7 @@ import { Component } from '../../../../core';
 import { RigidBody } from '../rigid-body';
 import { Eventify } from '../../../../core/event';
 import { IBaseConstraint } from '../../../spec/i-physics-constraint';
-import { createConstraint } from '../../instance';
+import { createConstraint } from '../../physics-selector';
 import { EConstraintType } from '../../physics-enum';
 
 /**
